# lib/api-zod/src/generated/types/createNotificationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNotificationBody {
  sourceId: string;
  kind: string;
  /** @nullable */
  occurredAt?: Date | null;
  title: string;
  /** @nullable */
  snippet?: string | null;
  /** @nullable */
  senderName?: string | null;
  /** @nullable */
  senderIdentifier?: string | null;
  /** @nullable */
  externalRef?: string | null;
}