# lib/api-zod/src/generated/types/createSourceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSourceBody {
  provider: string;
  accountIdentifier: string;
  /** @nullable */
  displayLabel?: string | null;
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  refreshToken?: string | null;
  /** @nullable */
  tokenExpiresAt?: Date | null;
}