# lib/api-zod/src/generated/types/privacyNotice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PrivacyNoticeDetails } from "./privacyNoticeDetails";

export interface PrivacyNotice {
  encryption: string;
  passwords: string;
  details: PrivacyNoticeDetails;
}