# lib/api-zod/src/generated/types/sourceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Source } from "./source";

export interface SourceListResponse {
  items: Source[];
}